scripts/requirements.txt
requests==2.26.0
pandas==1.3.3
python-dotenv==0.19.1
psycopg2==2.9.1
sqlalchemy==1.4.25